refactor to url.js
