+++ app/assets/src/components/PipelineSampleReport.jsx
@@ -32,11 +32,8 @@ import PhyloTreeChecks from "./views/phylo_tree/PhyloTreeChecks";
 import TaxonTreeVis from "./views/TaxonTreeVis";
 import LoadingLabel from "./ui/labels/LoadingLabel";
 import HoverActions from "./views/report/ReportTable/HoverActions";
-import {
-  getSampleReportInfo,
-  getSummaryContigCounts,
-  parseUrlParams
-} from "~/api";
+import { getSampleReportInfo, getSummaryContigCounts } from "~/api";
+import { parseUrlParams } from "~/helpers/url";
 import { pipelineVersionHasAssembly } from "./utils/sample";
 
 const DEFAULT_MIN_CONTIG_SIZE = 4;
